+++ main.py
@@ -1,24 +1,33 @@
 import cv2
 from bird_detect import detect_bird
 from detect_motion import detect_motion
-import socket
-import struct
-from PIL import Image
-import io
+import os
+import time
+from datetime import datetime
+import requests
 
-HOST, PORT = "localhost", 8000
+
 previous_frame = None
+img_counter = 0  # Counter to name saved images uniquely
+SAVE_DIR = "bird-detected"
+webhook_url = 'https://b25d-82-80-173-170.ngrok-free.app/detected'
+telegram_chat_id = '5669380497'
+
+# Create the directory if it doesn't exist
+if not os.path.exists(SAVE_DIR):
+    os.makedirs(SAVE_DIR)
 
 
-def send_image_to_pc(frame):
-    image_stream = io.BytesIO()
-    img = Image.fromarray(cv2.cvtColor(frame, cv2.COLOR_BGR2RGB))
-    img.save(image_stream, format='JPEG')
-    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-        s.connect((HOST, PORT))
-        image_data = image_stream.getvalue()
-        s.sendall(struct.pack('<L', len(image_data)))
-        s.sendall(image_data)
+def save_image_locally(frame):
+    global img_counter
+    # Get the current timestamp
+    timestamp = datetime.now().strftime('%Y-%m-%d_%H-%M-%S')
+
+    # Modify the save path to include the bird-detected directory and the timestamp
+    img_name = os.path.join(SAVE_DIR, f"detected_bird_{timestamp}_{img_counter}.jpg")
+    cv2.imwrite(img_name, frame)
+    print(f"Image saved as {img_name}")
+    img_counter += 1
 
 
 def open_camera():
@@ -33,6 +42,23 @@ def close_camera(cap):
     cv2.destroyAllWindows()
 
 
+def send_message_to_telegram_bot():
+    now = datetime.now()
+    formatted_time = now.strftime("%H:%M")
+    formatted_date = now.strftime("%d.%m.%Y")
+
+    message = f"A bird was detected at {formatted_time} on {formatted_date}"
+
+    payload = {
+        'url': webhook_url,
+        'chat_id': telegram_chat_id,
+        'text': message,
+    }
+
+    response = requests.post(webhook_url, json=payload)
+    return response
+
+
 def run(cap):
     ret, pre_frame = cap.read()
 
@@ -41,10 +67,14 @@ def run(cap):
 
         if detect_motion(current_frame, pre_frame) and detect_bird(current_frame):
             print("Motion detected, and a bird has been spotted!")
-            send_image_to_pc(current_frame)
-            user_input = input("Do you want to continue? (yes/no): ").strip().lower()
-            if user_input != 'yes':
-                break
+            save_image_locally(current_frame)
+
+            # msg to telegram
+            response = send_message_to_telegram_bot()
+            if response.status_code == 200:
+                print("message to farmer sent successfully")
+
+            time.sleep(3)  # Pause the program for 3 seconds
 
         pre_frame = current_frame.copy()
